Initial server implementation, very bare bones
Includes the rough outlines of how to register a journalist, mostly
sticking with how the Python server does it.

But I stopped because I don't have a data storage solution figured out
(don't want to use redis), and I want to spend a bit of time on types,
so we can have one type that encapsulates the key and intermediate
signature and doesn't require constant decoding, etc.